Fix: fix bug that enabled "-0"
Before: If you entered a decimal number and then used the negate
button to negate it, you could press the "del" button to delete
the decimals and decimal point so you end up with a negated 0 "-0"

example: 0.113 > negate > -0.113 > delete dec > -0. > delete dot >
            -0 (this shouldnt be allowed)

after: added function to check if screenB === "-0"
added this check into the deleteLatestInput switch case function
and if true, resetScreenB() will be run.